#405: Prevent overflow of uint64_t during multiplication, consume block resources after all transactions are applied

+++ libraries/chain/resources.cpp
@@ -68,7 +68,7 @@ void resource_meter::use_disk_storage( uint64_t bytes )
 
    if ( auto session = _session.lock() )
    {
-      uint128_t rc_cost = bytes * _resource_limit_data.disk_storage_cost();
+      uint128_t rc_cost = uint128_t( bytes ) * _resource_limit_data.disk_storage_cost();
       KOINOS_ASSERT( rc_cost <= std::numeric_limits< uint64_t >::max(), rc_overflow, "rc overflow" );
       session->use( rc_cost.convert_to< uint64_t >() );
    }
@@ -102,7 +102,7 @@ void resource_meter::use_network_bandwidth( uint64_t bytes )
 
    if ( auto session = _session.lock() )
    {
-      uint128_t rc_cost = bytes * _resource_limit_data.network_bandwidth_cost();
+      uint128_t rc_cost = uint128_t( bytes ) * _resource_limit_data.network_bandwidth_cost();
       KOINOS_ASSERT( rc_cost <= std::numeric_limits< uint64_t >::max(), rc_overflow, "rc overflow" );
       session->use( rc_cost.convert_to< uint64_t >() );
    }
@@ -136,7 +136,7 @@ void resource_meter::use_compute_bandwidth( uint64_t ticks )
 
    if ( auto session = _session.lock() )
    {
-      uint128_t rc_cost = ticks * _resource_limit_data.compute_bandwidth_cost();
+      uint128_t rc_cost = uint128_t( ticks ) * _resource_limit_data.compute_bandwidth_cost();
       KOINOS_ASSERT( rc_cost <= std::numeric_limits< uint64_t >::max(), rc_overflow, "rc overflow" );
       session->use( rc_cost.convert_to< uint64_t >() );
    }
